Modified Files Modified Entity files to edit log output Modified Status,Reward & ForeignExchangeCurrency repositories to take int values instead of long Modified Service classes to add CRUD operations

+++ src/main/java/com/fdmgroup/apmproject/repository/ForeignExchangeCurrencyRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.fdmgroup.apmproject.model.ForeignExchangeCurrency;
 
 @Repository
-public interface ForeignExchangeCurrencyRepository extends JpaRepository<ForeignExchangeCurrency, Long> {
+public interface ForeignExchangeCurrencyRepository extends JpaRepository<ForeignExchangeCurrency, Integer> {
 
 }
